[api/refacto] Move user to /core services
