+++ api/src/services/auth/authmanagement/authmanagement.hooks.ts
@@ -4,7 +4,7 @@ import { resolveExternal } from '@feathersjs/schema/lib'
 import { AuthenticationManagementService } from 'feathers-authentication-management/dist'
 import { iff, lowerCase, ContextFunction } from 'feathers-hooks-common'
 import { Application, HookContext } from '../../../declarations'
-import { userDispatchResolver } from '../user/user.resolver'
+import { userDispatchResolver } from '@/services/core/user/user.resolver'
 import { enforcePasswordPolicy } from './enforcePasswordPolicy.hook'
 
 const isAction =
