+++ api/src/services/core/user-group/user-group.model.ts
@@ -1,6 +1,6 @@
 import { Model, RelationMappings } from 'objection'
 import { GroupModel } from '../group/group.model'
-import { UserModel } from '../../auth/user/user.model'
+import { UserModel } from '../../core/user/user.model'
 
 /**
  * user-group objection Model
