+++ api/src/services/core/user/notifyUserOnUpdate.hooks.ts
@@ -1,6 +1,6 @@
-import { Application } from '../../../declarations'
+import { Application } from '@/declarations'
 import { HookContext } from '@feathersjs/feathers'
-import { authManagementSettings } from '../authmanagement/authmanagement.settings'
+import { authManagementSettings } from '@/services/auth/authmanagement/authmanagement.settings'
 import { User } from 'feathers-authentication-management'
 
 /**
